Clean/added tests, NRE check in APC
