+++ UnityProject/Assets/Scripts/Electricity/PoweredDevices/APC.cs
@@ -1,6 +1,8 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using System.Diagnostics;
+using System.Linq;
 using UnityEngine;
 using Mirror;
 
@@ -43,6 +45,24 @@ public class APC : NetworkBehaviour, ICheckedInteractable<HandApply>, INodeContr
 	{
 		voltageSync = newVoltage;
 		UpdateDisplay();
+	}
+
+	private void Start()
+	{
+		CheckListOfDevicesForNulls();
+	}
+
+	private void CheckListOfDevicesForNulls()
+	{
+		if (ConnectedDevices.Count == 0) return;
+		for (int i = ConnectedDevices.Count -1; i >= 0; i--)
+		{
+			if (ConnectedDevices[i] == null)
+			{
+				UnityEngine.Debug.Log($"{this.name} has a null value in {i}.");
+				ConnectedDevices.RemoveAt(i);
+			}
+		}
 	}
 
 	public override void OnStartServer()
@@ -140,7 +160,6 @@ public class APC : NetworkBehaviour, ICheckedInteractable<HandApply>, INodeContr
 
 		foreach (APCPoweredDevice Device in ConnectedDevices)
 		{
-			if(Device == null) continue;
 			Device.PowerNetworkUpdate(Voltages);
 			CalculatingResistance += (1 / Device.Resistance);
 		}
