--- UnityProject/Assets/Tests/LightSwitchTests.cs.meta
@@ -1,11 +1,0 @@
-fileFormatVersion: 2
-guid: 87837c8ddcf6dde47964619d30c4e5eb
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences: []
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
